# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AnalysisJobDTO": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AnalysisJobDTO",
        "type": "object"
      },
      "AnalysisROI": {
        "additionalProperties": false,
        "properties": {
          "coordinate_space": {
            "const": "original_px",
            "default": "original_px",
            "title": "Coordinate Space",
            "type": "string"
          },
          "invalid_rects": {
            "items": {
              "$ref": "#/components/schemas/InvalidPixelRegion"
            },
            "title": "Invalid Rects",
            "type": "array"
          },
          "revision": {
            "default": 1,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "source": {
            "default": "none",
            "enum": [
              "none",
              "manual",
              "detected"
            ],
            "title": "Source",
            "type": "string"
          },
          "valid_rect": {
            "$ref": "#/components/schemas/PixelRect"
          }
        },
        "required": [
          "valid_rect"
        ],
        "title": "AnalysisROI",
        "type": "object"
      },
      "ApiErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
      "ApiResponse_BoxSetDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoxSetDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[BoxSetDTO]",
        "type": "object"
      },
...
      "ApiResponse_JobDetailDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobDetailDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[JobDetailDTO]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__object__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[dict[str, object]]",
        "type": "object"
      },
      "ApiStatus": {
        "enum": [
          "success",
          "accepted",
          "error"
        ],
        "title": "ApiStatus",
        "type": "string"
      },
      "Body_createAnalysis": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          },
          "metadata_json": {
            "minLength": 2,
            "title": "Metadata Json",
            "type": "string"
          }
        },
        "required": [
          "files",
          "metadata_json"
        ],
        "title": "Body_createAnalysis",
        "type": "object"
      },
...
      "BoxSetDTO": {
        "additionalProperties": false,
        "properties": {
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/ROIBox"
            },
            "maxItems": 20,
            "title": "Boxes",
            "type": "array"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "image_id",
          "revision"
        ],
        "title": "BoxSetDTO",
        "type": "object"
      },
...
      "ImageAssetDTO": {
        "additionalProperties": false,
        "properties": {
          "analysis_roi": {
            "$ref": "#/components/schemas/AnalysisROI"
          },
          "bit_depth": {
            "exclusiveMinimum": 0.0,
            "title": "Bit Depth",
            "type": "integer"
          },
          "experiment_conditions": {
            "additionalProperties": true,
            "title": "Experiment Conditions",
            "type": "object"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "material_formula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Formula"
          },
          "material_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Name"
          },
          "original_download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Download Url"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          },
          "scale_nm_per_pixel": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Nm Per Pixel"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "image_id",
          "job_id",
          "filename",
          "sha256",
          "width",
          "height",
          "bit_depth",
          "sample_id",
          "analysis_roi"
        ],
        "title": "ImageAssetDTO",
        "type": "object"
      },
...
      "JobDetailDTO": {
        "additionalProperties": false,
        "properties": {
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageAssetDTO"
            },
            "title": "Images",
            "type": "array"
          },
          "job": {
            "$ref": "#/components/schemas/AnalysisJobDTO"
          },
          "partial_failures": {
            "items": {
              "$ref": "#/components/schemas/RunFailureDTO"
            },
            "title": "Partial Failures",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/SegmentationRunDTO"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "job"
        ],
        "title": "JobDetailDTO",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "CREATED",
          "VALIDATING",
          "READY_FOR_CONFIGURATION",
          "QUEUED",
          "PREPROCESSING",
          "SEGMENTING",
          "POSTPROCESSING",
          "QUALITY_CHECKING",
          "ANALYZING",
          "AGGREGATING",
          "COMPLETED",
          "COMPLETED_WITH_WARNINGS",
          "FAILED"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ROIBox": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "box_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Box Id"
          },
          "label": {
            "default": "",
            "maxLength": 120,
            "title": "Label",
            "type": "string"
          },
          "x1": {
            "minimum": 0.0,
            "title": "X1",
            "type": "integer"
          },
          "x2": {
            "exclusiveMinimum": 0.0,
            "title": "X2",
            "type": "integer"
          },
          "y1": {
            "minimum": 0.0,
            "title": "Y1",
            "type": "integer"
          },
          "y2": {
            "exclusiveMinimum": 0.0,
            "title": "Y2",
            "type": "integer"
          }
        },
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "ROIBox",
        "type": "object"
      },
...
      "ReplaceBoxesRequest": {
        "additionalProperties": false,
        "properties": {
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/ROIBox"
            },
            "maxItems": 20,
            "title": "Boxes",
            "type": "array"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "ReplaceBoxesRequest",
        "type": "object"
      },
...
      "RunFailureDTO": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "image_id",
          "model_id",
          "error_code"
        ],
        "title": "RunFailureDTO",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Traceable SEM nanoparticle analysis and evidence-backed query API",
    "title": "NanoLoop Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "operationId": "createAnalysis",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createAnalysis"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobDetailDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Create Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/api/v1/analyses/{job_id}": {
      "get": {
        "operationId": "getAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobDetailDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
...
    "/api/v1/analyses/{job_id}/images/{image_id}/boxes": {
      "get": {
        "operationId": "getBoxes",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BoxSetDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Get Boxes",
        "tags": [
          "boxes"
        ]
      },
      "put": {
        "operationId": "replaceBoxes",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceBoxesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BoxSetDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Replace Boxes",
        "tags": [
          "boxes"
        ]
      }
    },
...
  }
}